RSA: Move `test_signature_rsa_pkcs1_sign_output_buffer_len` to tests/.
It's only using the public API so it should be an integration test.
