+++ src/rsa/signing.rs
@@ -601,43 +601,3 @@ impl RsaKeyPair {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod tests {
-    // We intentionally avoid `use super::*` so that we are sure to use only
-    // the public API; this ensures that enough of the API is public.
-    use crate::{rand, signature};
-    use alloc::vec;
-
-    // `KeyPair::sign` requires that the output buffer is the same length as
-    // the public key modulus. Test what happens when it isn't the same length.
-    #[test]
-    fn test_signature_rsa_pkcs1_sign_output_buffer_len() {
-        // Sign the message "hello, world", using PKCS#1 v1.5 padding and the
-        // SHA256 digest algorithm.
-        const MESSAGE: &[u8] = b"hello, world";
-        let rng = rand::SystemRandom::new();
-
-        const PRIVATE_KEY_DER: &[u8] = include_bytes!("signature_rsa_example_private_key.der");
-        let key_pair = signature::RsaKeyPair::from_der(PRIVATE_KEY_DER).unwrap();
-
-        // The output buffer is one byte too short.
-        let mut signature = vec![0; key_pair.public_modulus_len() - 1];
-
-        assert!(key_pair
-            .sign(&signature::RSA_PKCS1_SHA256, &rng, MESSAGE, &mut signature)
-            .is_err());
-
-        // The output buffer is the right length.
-        signature.push(0);
-        assert!(key_pair
-            .sign(&signature::RSA_PKCS1_SHA256, &rng, MESSAGE, &mut signature)
-            .is_ok());
-
-        // The output buffer is one byte too long.
-        signature.push(0);
-        assert!(key_pair
-            .sign(&signature::RSA_PKCS1_SHA256, &rng, MESSAGE, &mut signature)
-            .is_err());
-    }
-}
